Add `channel list' command.
This command acts like `show' except that it does not take
any arguments and simply prints the entire list.  `channel show'
will now show a single channel if the identifier is specified.

`channel show 1' for example.  Keep in mind this is still a bit
buggy for invalid channels.  I'm not sure how to verify a DBUS
path is valid before calling it.