[agent] Make subscriptions work directly with sources.
The subscription system had been receiving information directly
from the channels.  This caused extra contention and would be a
source of potential deadlocks.  This design change makes sources
work directly with subscriptions.  A channel may be added to a
subscription and all of its sources, even newly added or removed
sources, will be tracked in the subscription.